requirements.txt
# FastAPI 相关
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# HTTP 请求
requests>=2.31.0

# LangChain 相关
langchain>=0.1.0
langchain-core>=0.1.0

# 图像处理（如果需要）
pillow>=10.0.0

# PyTorch 和 NumPy（固定版本以避免兼容性问题）
# 注意：PyTorch 需要从 CPU 专用源安装：
# pip install torch==2.3.1 --index-url https://download.pytorch.org/whl/cpu
# NumPy 2.x 与 PyTorch 2.3.1 不兼容，必须使用 1.x 版本
numpy<2.0,>=1.24.0







